Converted stored procedures to nHibernate
git-svn-id: https://svn.esynctraining.com/svn/esynctraining@3492 41667c05-fd31-8a45-9ea7-b4ae18c5af78

+++ EdugameCloud.Core/Domain/DTO/TestDTO.cs
@@ -53,7 +53,7 @@
         /// Gets or sets the passing score.
         /// </summary>
         [DataMember]
-        public float? passingScore { get; set; }
+        public decimal? passingScore { get; set; }
 
         /// <summary>
         /// Gets or sets the score format.
