+++ EdugameCloud.Core/Domain/DTO/TestPlayerDTO.cs
@@ -75,7 +75,7 @@
         /// Gets or sets the passing score.
         /// </summary>
         [DataMember]
-        public float? passingScore { get; set; }
+        public decimal? passingScore { get; set; }
 
         #endregion
     }
